--- a/lib/generated/l10n/zulip_localizations_ar.dart
+++ b/lib/generated/l10n/zulip_localizations_ar.dart
@@ -644,6 +644,12 @@ class ZulipLocalizationsAr extends ZulipLocalizations {
   String get pollWidgetOptionsMissing => 'This poll has no options yet.';
 
   @override
+  String get experimentalFeatureSettingsPageTitle => 'Experimental features';
+
+  @override
+  String get experimentalFeatureSettingsWarning => 'These options enable features which are still under development and not ready. They may not work, and may cause issues in other areas of the app.\n\nThe purpose of these settings is for experimentation by people working on developing Zulip.';
+
+  @override
   String get errorNotificationOpenTitle => 'Failed to open notification';
 
   @override