--- a/lib/model/settings.dart
+++ b/lib/model/settings.dart
@@ -60,11 +60,22 @@ class GlobalSettingsStore extends ChangeNotifier {
   /// A cache of the [GlobalSettingsData] singleton in the underlying data store.
   GlobalSettingsData _data;
 
+  Future<void> _update(GlobalSettingsCompanion data) async {
+    await _backend.doUpdateGlobalSettings(data);
+    _data = _data.copyWithCompanion(data);
+    notifyListeners();
+  }
+
   /// The user's choice of [ThemeSetting];
   /// null means the device-level choice of theme.
   ///
   /// See also [setThemeSetting].
   ThemeSetting? get themeSetting => _data.themeSetting;
+
+  /// Set [themeSetting], persistently for future runs of the app.
+  Future<void> setThemeSetting(ThemeSetting? value) async {
+    await _update(GlobalSettingsCompanion(themeSetting: Value(value)));
+  }
 
   /// The user's choice of [BrowserPreference];
   /// null means use our default choice.
@@ -73,6 +84,11 @@ class GlobalSettingsStore extends ChangeNotifier {
   ///
   /// See also [setBrowserPreference].
   BrowserPreference? get browserPreference => _data.browserPreference;
+
+  /// Set [browserPreference], persistently for future runs of the app.
+  Future<void> setBrowserPreference(BrowserPreference? value) async {
+    await _update(GlobalSettingsCompanion(browserPreference: Value(value)));
+  }
 
   /// The value of [BrowserPreference] to use:
   /// the user's choice [browserPreference] if any, else our default.
@@ -112,21 +128,5 @@ class GlobalSettingsStore extends ChangeNotifier {
       case BrowserPreference.external:
         return UrlLaunchMode.externalApplication;
     }
-  }
-
-  Future<void> _update(GlobalSettingsCompanion data) async {
-    await _backend.doUpdateGlobalSettings(data);
-    _data = _data.copyWithCompanion(data);
-    notifyListeners();
-  }
-
-  /// Set [themeSetting], persistently for future runs of the app.
-  Future<void> setThemeSetting(ThemeSetting? value) async {
-    await _update(GlobalSettingsCompanion(themeSetting: Value(value)));
-  }
-
-  /// Set [browserPreference], persistently for future runs of the app.
-  Future<void> setBrowserPreference(BrowserPreference? value) async {
-    await _update(GlobalSettingsCompanion(browserPreference: Value(value)));
   }
 }