Fixes #1409.

(**Note NOMERGE demo commit** at the end. That illustrates how to use this new facility, and should be removed before merging.)

This introduces a system for experimental feature flags which the user can enable or disable in settings.

One thing I'd like to avoid with such flags is having to do a database migration each time we add or remove one. (Those are fairly heavyweight with Drift's migration system, as each schema version produces several additional copies of our schema in various generated files. One aspect I appreciate in Django's migration system, as seen in the Zulip server, is that it avoids that sort of overhead.)

In order to do that, this puts the new flags as rows in a table, effectively a key-value map. The assumptions that make that table work don't really use the fact that these are specifically being used as experimental feature flags; so we set it up to generically accommodate any sort of boolean global setting, in case we add others in the future.

When no experimental feature flag currently exists in the app, the settings UI for them doesn't appear. That's true in the version of this PR intended for merge, without the NOMERGE commit at the end.

That NOMERGE commit therefore enables manually testing this UI. It also serves to demonstrate adding a new feature flag in this system, which involves a very small amount of code.


## Selected commit messages

#### f036feaa9 log: Introduce profilePrint; use it for timing initial fetch


#### 285414cc4 db: Add profile-mode timings of load at startup

As we add more features using the database and more queries we make
from it here at the app's startup, this will help us check that we
aren't unduly delaying startup.

Results on my Pixel 8, with 2 account records:

db load time 91.9ms total: 1.5ms init, 85.4ms settings, 5.0ms accounts
db load time 87.8ms total: 1.7ms init, 78.1ms settings, 8.0ms accounts
db load time 93.3ms total: 1.2ms init, 83.7ms settings, 8.5ms accounts
db load time 78.1ms total: 1.4ms init, 71.8ms settings, 4.8ms accounts
db load time 86.9ms total: 1.2ms init, 77.8ms settings, 7.9ms accounts


#### e50b68dec settings: Make general bool global settings, so as to add without migrations

This new table to read looks to add something like 5-8ms to the app's
startup time.  Not nothing; but not a big cost compared to the time
we spend loading server data right afterward.

Specifically, results on my Pixel 8 (compare to those in the
previous commit):

db load time 117.9ms total: 15.2ms init, 89.2ms settings, 5.1ms bool-settings, 8.3ms accounts
db load time 90.9ms total: 1.3ms init, 78.5ms settings, 8.3ms bool-settings, 2.8ms accounts
db load time 87.2ms total: 1.4ms init, 71.8ms settings, 5.9ms bool-settings, 8.1ms accounts
db load time 85.7ms total: 1.2ms init, 72.8ms settings, 4.9ms bool-settings, 6.8ms accounts
db load time 91.3ms total: 1.4ms init, 80.0ms settings, 7.5ms bool-settings, 2.5ms accounts
db load time 83.8ms total: 1.1ms init, 70.0ms settings, 5.0ms bool-settings, 7.7ms accounts


#### e36ed8eb0 settings: Add the concept of experimental feature flags

Fixes #1409.


#### 216eecdfa NOMERGE demo flag: renderKatex

This demonstrates what's involved in introducing a new
experimental feature flag.