@gnprice 2025-03-19
Screenshots:

| Main settings page | New settings sub-page |
| --- | --- |
| ![image](https://github.com/user-attachments/assets/eca2e968-533b-4d69-9d65-86ee9bf0e9fe) | ![image](https://github.com/user-attachments/assets/0137b561-d950-4485-a242-e0cf578e6d44) |