@gnprice 2025-03-19
One more thing before this should be merged: #1416 should have adjusted `AppDatabase._dropAndCreateAll` to reflect the change it made to the `onCreate` in `migration`, making it insert the GlobalSettings row.

Without that, if you switch to a branch (like this one) with a later schema version, then back to schema version 5 — only really possible as a developer — you get this error at startup:
```
E/flutter (15620): [ERROR:flutter/runtime/dart_vm_initializer.cc(40)] Unhandled Exception: Bad state: No element
E/flutter (15620): #0      List.single (dart:core-patch/growable_array.dart:358:22)
E/flutter (15620): #1      Selectable.getSingle (package:drift/src/runtime/query_builder/statements/query.dart:239:26)
E/flutter (15620): <asynchronous suspension>
E/flutter (15620): #2      AppDatabase.getGlobalSettings (package:zulip/model/database.dart:186:12)
E/flutter (15620): <asynchronous suspension>
E/flutter (15620): #3      LiveGlobalStore.load (package:zulip/model/store.dart:886:28)
E/flutter (15620): <asynchronous suspension>
E/flutter (15620): #4      _GlobalStoreWidgetState.initState.<anonymous closure> (package:zulip/widgets/store.dart:89:21)
E/flutter (15620): <asynchronous suspension>
```
and the app fails to get past the loading screen.

To unwedge it, you can switch to a version before schema version 5 (e.g., to this week's v0.0.27 release) and then forward again; or uninstall the app and reinstall.

I'll fix that, sending a separate PR. That should be merged before this one is, to reduce the window of history that can trigger the bug.

I don't think any of the changes in this PR will need to be revised for that, though. So this remains ready for review.