@gnprice 2025-03-21
Thanks for the careful review! Pushed a revision; PTAL.

Also sent #1428 to handle the downgrades issue mentioned above. Before merge, this should wait for that, and also have its NOMERGE commit dropped.