@gnprice 2025-03-24
Thanks! And #1428 is now merged, so merging this.

FTR here's that NOMERGE demo commit which I'm now dropping:
b3512ed00 NOMERGE demo flag: renderKatex

<details>
<summary>
<code>git log --stat -p -1</code>
</summary>

```diff
commit b3512ed0039fa7e057ec80c069403244c4933a3d
Author: Greg Price <greg@zulip.com>
Date:   Sun Mar 16 22:16:57 2025 -0700

    NOMERGE demo flag: renderKatex
    
    This demonstrates what's involved in introducing a new
    experimental feature flag.
---
 lib/model/settings.dart  |  2 ++
 lib/widgets/content.dart | 37 +++++++++++++++++++++++++++----------
 2 files changed, 29 insertions(+), 10 deletions(-)

diff --git lib/model/settings.dart lib/model/settings.dart
index dc1bdc524..d7b7161e8 100644
--- lib/model/settings.dart
+++ lib/model/settings.dart
@@ -110,6 +110,8 @@ enum BoolGlobalSetting {
   /// (Having one stable value in this enum is also handy for tests.)
   placeholderIgnore(GlobalSettingType.placeholder, false),
 
+  renderKatex(GlobalSettingType.experimentalFeatureFlag, false),
+
   // Former settings which might exist in the database,
   // whose names should therefore not be reused:
   // (this list is empty so far)
diff --git lib/widgets/content.dart lib/widgets/content.dart
index c94960282..20e39e866 100644
--- lib/widgets/content.dart
+++ lib/widgets/content.dart
@@ -15,6 +15,7 @@ import '../model/avatar_url.dart';
 import '../model/binding.dart';
 import '../model/content.dart';
 import '../model/internal_link.dart';
+import '../model/settings.dart';
 import 'code_block.dart';
 import 'dialog.dart';
 import 'icons.dart';
@@ -831,11 +832,19 @@ class MathBlock extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final contentTheme = ContentTheme.of(context);
-    return _CodeBlockContainer(
-      borderColor: contentTheme.colorMathBlockBorder,
-      child: Text.rich(TextSpan(
-        style: contentTheme.codeBlockTextStyles.plain,
-        children: [TextSpan(text: node.texSource)])));
+    final globalSettings = GlobalStoreWidget.settingsOf(context);
+
+    final renderKatex = globalSettings.getBool(BoolGlobalSetting.renderKatex);
+    if (!renderKatex) {
+      return _CodeBlockContainer(
+        borderColor: contentTheme.colorMathBlockBorder,
+        child: Text.rich(TextSpan(
+          style: contentTheme.codeBlockTextStyles.plain,
+          children: [TextSpan(text: node.texSource)])));
+    }
+
+    return Text(style: contentTheme.textStyleError,
+      '(error: KaTeX unimplemented)'); // TODO(#1408)
   }
 }
 
@@ -1147,11 +1156,19 @@ class _InlineContentBuilder {
           child: MessageImageEmoji(node: node));
 
       case MathInlineNode():
-        return TextSpan(
-          style: widget.style
-            .merge(ContentTheme.of(_context!).textStyleInlineMath)
-            .apply(fontSizeFactor: kInlineCodeFontSizeFactor),
-          children: [TextSpan(text: node.texSource)]);
+        final contentTheme = ContentTheme.of(_context!);
+        final globalSettings = GlobalStoreWidget.settingsOf(_context!);
+        final renderKatex = globalSettings.getBool(BoolGlobalSetting.renderKatex);
+        if (!renderKatex) {
+          return TextSpan(
+            style: widget.style
+              .merge(ContentTheme.of(_context!).textStyleInlineMath)
+              .apply(fontSizeFactor: kInlineCodeFontSizeFactor),
+            children: [TextSpan(text: node.texSource)]);
+        }
+
+        return TextSpan(style: contentTheme.textStyleError,
+          text: '(error: KaTeX unimplemented)'); // TODO(#1408)
 
       case GlobalTimeNode():
         return WidgetSpan(alignment: PlaceholderAlignment.middle,
```

</details>